Struct google_analytics3::Account

pub struct Account {
    pub kind: Option<String>,
    pub name: Option<String>,
    pub created: Option<String>,
    pub updated: Option<String>,
    pub child_link: Option<AccountChildLink>,
    pub starred: Option<bool>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub permissions: Option<AccountPermissions>,
}

JSON template for Analytics account entry.

This type is not used in any activity, and only used as part of another schema.

Fields

kind

Resource type for Analytics account.

name

Account name.

created

Time the account was created.

updated

Time the account was last modified.

child_link

Child link for an account entry. Points to the list of web properties for this account.

starred

Indicates whether this account is starred or not.

id

Account ID.

self_link

Link for this account.

permissions

Permissions the user has for this account.