Struct google_analytics3::AccountSummaries

pub struct AccountSummaries {
    pub username: Option<String>,
    pub kind: Option<String>,
    pub items: Option<Vec<AccountSummary>>,
    pub items_per_page: Option<i32>,
    pub previous_link: Option<String>,
    pub start_index: Option<i32>,
    pub next_link: Option<String>,
    pub total_results: Option<i32>,
}

An AccountSummary collection lists a summary of accounts, properties and views (profiles) to which the user has access. Each resource in the collection corresponds to a single AccountSummary.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

account summaries list management (response)

Fields
username: Option<String>

Email ID of the authenticated user

kind: Option<String>

Collection type.

items: Option<Vec<AccountSummary>>

A list of AccountSummaries.

items_per_page: Option<i32>

The maximum number of resources the response can contain, regardless of the actual number of resources returned. Its value ranges from 1 to 1000 with a value of 1000 by default, or otherwise specified by the max-results query parameter.

previous_link: Option<String>

Link to previous page for this AccountSummary collection.

start_index: Option<i32>

The starting index of the resources, which is 1 by default or otherwise specified by the start-index query parameter.

next_link: Option<String>

Link to next page for this AccountSummary collection.

total_results: Option<i32>

The total number of results for the query, regardless of the number of results in the response.