Struct google_analytics3::EntityUserLink

pub struct EntityUserLink {
    pub kind: Option<String>,
    pub user_ref: Option<UserRef>,
    pub permissions: Option<EntityUserLinkPermissions>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub entity: Option<EntityUserLinkEntity>,
}

JSON template for an Analytics Entity-User Link. Returns permissions that a user has for an entity.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

webproperty user links insert management (request|response)

profile user links insert management (request|response)

profile user links update management (request|response)

account user links update management (request|response)

webproperty user links update management (request|response)

account user links insert management (request|response)

Fields
kind: Option<String>

Resource type for entity user link.

user_ref: Option<UserRef>

User reference.

permissions: Option<EntityUserLinkPermissions>

Permissions the user has for this entity.

id: Option<String>

Entity user link ID

self_link: Option<String>

Self link for this resource.

entity: Option<EntityUserLinkEntity>

Entity for this link. It can be an account, a web property, or a view (profile).