Struct google_analytics3::Filter

pub struct Filter {
    pub kind: Option<String>,
    pub name: Option<String>,
    pub created: Option<String>,
    pub updated: Option<String>,
    pub advanced_details: Option<FilterAdvancedDetails>,
    pub lowercase_details: Option<FilterLowercaseDetails>,
    pub parent_link: Option<FilterParentLink>,
    pub exclude_details: Option<FilterExpression>,
    pub include_details: Option<FilterExpression>,
    pub uppercase_details: Option<FilterUppercaseDetails>,
    pub account_id: Option<String>,
    pub type_: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub search_and_replace_details: Option<FilterSearchAndReplaceDetails>,
}

JSON template for an Analytics account filter.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

filters delete management (response)

filters patch management (request|response)

filters update management (request|response)

filters get management (response)

filters insert management (request|response)

Fields
kind: Option<String>

Resource type for Analytics filter.

name: Option<String>

Name of this filter.

created: Option<String>

Time this filter was created.

updated: Option<String>

Time this filter was last modified.

advanced_details: Option<FilterAdvancedDetails>

Details for the filter of the type ADVANCED.

lowercase_details: Option<FilterLowercaseDetails>

Details for the filter of the type LOWER.

parent_link: Option<FilterParentLink>

Parent link for this filter. Points to the account to which this filter belongs.

exclude_details: Option<FilterExpression>

Details for the filter of the type EXCLUDE.

include_details: Option<FilterExpression>

Details for the filter of the type INCLUDE.

uppercase_details: Option<FilterUppercaseDetails>

Details for the filter of the type UPPER.

account_id: Option<String>

Account ID to which this filter belongs.

type_: Option<String>

Type of this filter. Possible values are INCLUDE, EXCLUDE, LOWERCASE, UPPERCASE, SEARCH_AND_REPLACE and ADVANCED.

id: Option<String>

Filter ID.

self_link: Option<String>

Link for this filter.

search_and_replace_details: Option<FilterSearchAndReplaceDetails>

Details for the filter of the type SEARCH_AND_REPLACE.