Struct google_analytics3::FilterAdvancedDetails

pub struct FilterAdvancedDetails {
    pub field_a: Option<String>,
    pub override_output_field: Option<bool>,
    pub field_a_required: Option<bool>,
    pub output_constructor: Option<String>,
    pub field_b_required: Option<bool>,
    pub case_sensitive: Option<bool>,
    pub field_b: Option<String>,
    pub extract_b: Option<String>,
    pub extract_a: Option<String>,
    pub output_to_field: Option<String>,
}

Details for the filter of the type ADVANCED.

This type is not used in any activity, and only used as part of another schema.

Fields

field_a

Field A.

override_output_field

Indicates if the existing value of the output field, if any, should be overridden by the output expression.

field_a_required

Indicates if field A is required to match.

output_constructor

Expression used to construct the output value.

field_b_required

Indicates if field B is required to match.

case_sensitive

Indicates if the filter expressions are case sensitive.

field_b

Field B.

extract_b

Expression to extract from field B.

extract_a

Expression to extract from field A.

output_to_field

Output field.