Struct google_analytics3::Goal

pub struct Goal {
    pub updated: Option<String>,
    pub internal_web_property_id: Option<String>,
    pub web_property_id: Option<String>,
    pub active: Option<bool>,
    pub id: Option<String>,
    pub account_id: Option<String>,
    pub kind: Option<String>,
    pub visit_time_on_site_details: Option<GoalVisitTimeOnSiteDetails>,
    pub name: Option<String>,
    pub created: Option<String>,
    pub type_: Option<String>,
    pub visit_num_pages_details: Option<GoalVisitNumPagesDetails>,
    pub value: Option<f32>,
    pub event_details: Option<GoalEventDetails>,
    pub profile_id: Option<String>,
    pub parent_link: Option<GoalParentLink>,
    pub url_destination_details: Option<GoalUrlDestinationDetails>,
    pub self_link: Option<String>,
}

JSON template for Analytics goal resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

goals get management (response)

goals update management (request|response)

goals patch management (request|response)

goals insert management (request|response)

Fields

updated

Time this goal was last modified.

internal_web_property_id

Internal ID for the web property to which this goal belongs.

web_property_id

Web property ID to which this goal belongs. The web property ID is of the form UA-XXXXX-YY.

active

Determines whether this goal is active.

id

Goal ID.

account_id

Account ID to which this goal belongs.

kind

Resource type for an Analytics goal.

visit_time_on_site_details

Details for the goal of the type VISIT_TIME_ON_SITE.

name

Goal name.

created

Time this goal was created.

type_

Goal type. Possible values are URL_DESTINATION, VISIT_TIME_ON_SITE, VISIT_NUM_PAGES, AND EVENT.

visit_num_pages_details

Details for the goal of the type VISIT_NUM_PAGES.

value

Goal value.

event_details

Details for the goal of the type EVENT.

profile_id

View (Profile) ID to which this goal belongs.

parent_link

Parent link for a goal. Points to the view (profile) to which this goal belongs.

url_destination_details

Details for the goal of the type URL_DESTINATION.

self_link

Link for this goal.