Struct google-analytics3::ManagementMethodsBuilder

pub struct ManagementMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on management resources. It is not used directly, but through the Analytics hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate serde;
extern crate "google-analytics3" as analytics3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use analytics3::Analytics;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Analytics::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `account_summaries_list(...)`, `account_user_links_delete(...)`, `account_user_links_insert(...)`, `account_user_links_list(...)`, `account_user_links_update(...)`, `accounts_list(...)`, `custom_data_sources_list(...)`, `custom_dimensions_get(...)`, `custom_dimensions_insert(...)`, `custom_dimensions_list(...)`, `custom_dimensions_patch(...)`, `custom_dimensions_update(...)`, `custom_metrics_get(...)`, `custom_metrics_insert(...)`, `custom_metrics_list(...)`, `custom_metrics_patch(...)`, `custom_metrics_update(...)`, `experiments_delete(...)`, `experiments_get(...)`, `experiments_insert(...)`, `experiments_list(...)`, `experiments_patch(...)`, `experiments_update(...)`, `filters_delete(...)`, `filters_get(...)`, `filters_insert(...)`, `filters_list(...)`, `filters_patch(...)`, `filters_update(...)`, `goals_get(...)`, `goals_insert(...)`, `goals_list(...)`, `goals_patch(...)`, `goals_update(...)`, `profile_filter_links_delete(...)`, `profile_filter_links_get(...)`, `profile_filter_links_insert(...)`, `profile_filter_links_list(...)`, `profile_filter_links_patch(...)`, `profile_filter_links_update(...)`, `profile_user_links_delete(...)`, `profile_user_links_insert(...)`, `profile_user_links_list(...)`, `profile_user_links_update(...)`, `profiles_delete(...)`, `profiles_get(...)`, `profiles_insert(...)`, `profiles_list(...)`, `profiles_patch(...)`, `profiles_update(...)`, `segments_list(...)`, `unsampled_reports_get(...)`, `unsampled_reports_insert(...)`, `unsampled_reports_list(...)`, `uploads_delete_upload_data(...)`, `uploads_get(...)`, `uploads_list(...)`, `uploads_upload_data(...)`, `web_property_ad_words_links_delete(...)`, `web_property_ad_words_links_get(...)`, `web_property_ad_words_links_insert(...)`, `web_property_ad_words_links_list(...)`, `web_property_ad_words_links_patch(...)`, `web_property_ad_words_links_update(...)`, `webproperties_get(...)`, `webproperties_insert(...)`, `webproperties_list(...)`, `webproperties_patch(...)`, `webproperties_update(...)`, `webproperty_user_links_delete(...)`, `webproperty_user_links_insert(...)`, `webproperty_user_links_list(...)` and `webproperty_user_links_update(...)`
// to build up your call.
let rb = hub.management();

Methods

impl<'a, C, NC, A> ManagementMethodsBuilder<'a, C, NC, A>

fn webproperties_insert(&self, request: &Webproperty, account_id: &str) -> ManagementWebpropertyInsertCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Create a new property if the account has fewer than 20 properties. Web properties are visible in the Google Analytics interface only if they have at least one profile.

fn profiles_get(&self, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementProfileGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets a view (profile) to which the user has access.

fn accounts_list(&self) -> ManagementAccountListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists all accounts to which the user has access.

fn profile_filter_links_patch(&self, request: &ProfileFilterLink, account_id: &str, web_property_id: &str, profile_id: &str, link_id: &str) -> ManagementProfileFilterLinkPatchCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Update an existing profile filter link. This method supports patch semantics.

fn webproperty_user_links_delete(&self, account_id: &str, web_property_id: &str, link_id: &str) -> ManagementWebpropertyUserLinkDeleteCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Removes a user from the given web property.

fn profile_user_links_delete(&self, account_id: &str, web_property_id: &str, profile_id: &str, link_id: &str) -> ManagementProfileUserLinkDeleteCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Removes a user from the given view (profile).

fn profile_user_links_update(&self, request: &EntityUserLink, account_id: &str, web_property_id: &str, profile_id: &str, link_id: &str) -> ManagementProfileUserLinkUpdateCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates permissions for an existing user on the given view (profile).

fn filters_insert(&self, request: &Filter, account_id: &str) -> ManagementFilterInsertCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Create a new filter.

fn account_user_links_update(&self, request: &EntityUserLink, account_id: &str, link_id: &str) -> ManagementAccountUserLinkUpdateCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates permissions for an existing user on the given account.

fn webproperties_update(&self, request: &Webproperty, account_id: &str, web_property_id: &str) -> ManagementWebpropertyUpdateCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an existing web property.

fn webproperty_user_links_update(&self, request: &EntityUserLink, account_id: &str, web_property_id: &str, link_id: &str) -> ManagementWebpropertyUserLinkUpdateCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates permissions for an existing user on the given web property.

fn unsampled_reports_insert(&self, request: &UnsampledReport, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementUnsampledReportInsertCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Create a new unsampled report.

fn custom_metrics_get(&self, account_id: &str, web_property_id: &str, custom_metric_id: &str) -> ManagementCustomMetricGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Get a custom metric to which the user has access.

fn uploads_get(&self, account_id: &str, web_property_id: &str, custom_data_source_id: &str, upload_id: &str) -> ManagementUploadGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

List uploads to which the user has access.

fn web_property_ad_words_links_get(&self, account_id: &str, web_property_id: &str, web_property_ad_words_link_id: &str) -> ManagementWebPropertyAdWordsLinkGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns a web property-AdWords link to which the user has access.

fn webproperty_user_links_list(&self, account_id: &str, web_property_id: &str) -> ManagementWebpropertyUserLinkListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists webProperty-user links for a given web property.

fn profile_filter_links_update(&self, request: &ProfileFilterLink, account_id: &str, web_property_id: &str, profile_id: &str, link_id: &str) -> ManagementProfileFilterLinkUpdateCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Update an existing profile filter link.

fn experiments_get(&self, account_id: &str, web_property_id: &str, profile_id: &str, experiment_id: &str) -> ManagementExperimentGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns an experiment to which the user has access.

fn custom_dimensions_update(&self, request: &CustomDimension, account_id: &str, web_property_id: &str, custom_dimension_id: &str) -> ManagementCustomDimensionUpdateCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an existing custom dimension.

fn unsampled_reports_get(&self, account_id: &str, web_property_id: &str, profile_id: &str, unsampled_report_id: &str) -> ManagementUnsampledReportGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns a single unsampled report.

fn profile_filter_links_insert(&self, request: &ProfileFilterLink, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementProfileFilterLinkInsertCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Create a new profile filter link.

fn filters_update(&self, request: &Filter, account_id: &str, filter_id: &str) -> ManagementFilterUpdateCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an existing filter.

fn profile_filter_links_delete(&self, account_id: &str, web_property_id: &str, profile_id: &str, link_id: &str) -> ManagementProfileFilterLinkDeleteCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Delete a profile filter link.

fn segments_list(&self) -> ManagementSegmentListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists segments to which the user has access.

fn profiles_delete(&self, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementProfileDeleteCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes a view (profile).

fn goals_patch(&self, request: &Goal, account_id: &str, web_property_id: &str, profile_id: &str, goal_id: &str) -> ManagementGoalPatchCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an existing view (profile). This method supports patch semantics.

fn custom_dimensions_patch(&self, request: &CustomDimension, account_id: &str, web_property_id: &str, custom_dimension_id: &str) -> ManagementCustomDimensionPatchCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an existing custom dimension. This method supports patch semantics.

fn experiments_patch(&self, request: &Experiment, account_id: &str, web_property_id: &str, profile_id: &str, experiment_id: &str) -> ManagementExperimentPatchCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Update an existing experiment. This method supports patch semantics.

fn experiments_list(&self, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementExperimentListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists experiments to which the user has access.

fn profiles_list(&self, account_id: &str, web_property_id: &str) -> ManagementProfileListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists views (profiles) to which the user has access.

fn goals_list(&self, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementGoalListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists goals to which the user has access.

fn goals_insert(&self, request: &Goal, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementGoalInsertCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Create a new goal.

fn custom_metrics_patch(&self, request: &CustomMetric, account_id: &str, web_property_id: &str, custom_metric_id: &str) -> ManagementCustomMetricPatchCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an existing custom metric. This method supports patch semantics.

fn uploads_upload_data(&self, account_id: &str, web_property_id: &str, custom_data_source_id: &str) -> ManagementUploadUploadDataCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Upload data for a custom data source.

fn account_user_links_insert(&self, request: &EntityUserLink, account_id: &str) -> ManagementAccountUserLinkInsertCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Adds a new user to the given account.

fn webproperty_user_links_insert(&self, request: &EntityUserLink, account_id: &str, web_property_id: &str) -> ManagementWebpropertyUserLinkInsertCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Adds a new user to the given web property.

fn experiments_delete(&self, account_id: &str, web_property_id: &str, profile_id: &str, experiment_id: &str) -> ManagementExperimentDeleteCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Delete an experiment.

fn profiles_patch(&self, request: &Profile, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementProfilePatchCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an existing view (profile). This method supports patch semantics.

fn filters_get(&self, account_id: &str, filter_id: &str) -> ManagementFilterGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns a filters to which the user has access.

fn web_property_ad_words_links_list(&self, account_id: &str, web_property_id: &str) -> ManagementWebPropertyAdWordsLinkListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists webProperty-AdWords links for a given web property.

fn account_summaries_list(&self) -> ManagementAccountSummaryListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists account summaries (lightweight tree comprised of accounts/properties/profiles) to which the user has access.

fn custom_dimensions_list(&self, account_id: &str, web_property_id: &str) -> ManagementCustomDimensionListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists custom dimensions to which the user has access.

fn uploads_list(&self, account_id: &str, web_property_id: &str, custom_data_source_id: &str) -> ManagementUploadListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

List uploads to which the user has access.

fn profile_user_links_list(&self, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementProfileUserLinkListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists profile-user links for a given view (profile).

fn account_user_links_list(&self, account_id: &str) -> ManagementAccountUserLinkListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists account-user links for a given account.

fn custom_metrics_update(&self, request: &CustomMetric, account_id: &str, web_property_id: &str, custom_metric_id: &str) -> ManagementCustomMetricUpdateCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an existing custom metric.

fn uploads_delete_upload_data(&self, request: &AnalyticsDataimportDeleteUploadDataRequest, account_id: &str, web_property_id: &str, custom_data_source_id: &str) -> ManagementUploadDeleteUploadDataCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Delete data associated with a previous upload.

fn experiments_update(&self, request: &Experiment, account_id: &str, web_property_id: &str, profile_id: &str, experiment_id: &str) -> ManagementExperimentUpdateCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Update an existing experiment.

fn goals_update(&self, request: &Goal, account_id: &str, web_property_id: &str, profile_id: &str, goal_id: &str) -> ManagementGoalUpdateCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an existing view (profile).

fn custom_data_sources_list(&self, account_id: &str, web_property_id: &str) -> ManagementCustomDataSourceListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

List custom data sources to which the user has access.

fn custom_metrics_list(&self, account_id: &str, web_property_id: &str) -> ManagementCustomMetricListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists custom metrics to which the user has access.

fn unsampled_reports_list(&self, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementUnsampledReportListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists unsampled reports to which the user has access.

fn webproperties_get(&self, account_id: &str, web_property_id: &str) -> ManagementWebpropertyGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets a web property to which the user has access.

fn experiments_insert(&self, request: &Experiment, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementExperimentInsertCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Create a new experiment.

fn custom_dimensions_get(&self, account_id: &str, web_property_id: &str, custom_dimension_id: &str) -> ManagementCustomDimensionGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Get a custom dimension to which the user has access.

fn profile_filter_links_list(&self, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementProfileFilterLinkListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists all profile filter links for a profile.

fn webproperties_patch(&self, request: &Webproperty, account_id: &str, web_property_id: &str) -> ManagementWebpropertyPatchCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an existing web property. This method supports patch semantics.

fn custom_dimensions_insert(&self, request: &CustomDimension, account_id: &str, web_property_id: &str) -> ManagementCustomDimensionInsertCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Create a new custom dimension.

fn custom_metrics_insert(&self, request: &CustomMetric, account_id: &str, web_property_id: &str) -> ManagementCustomMetricInsertCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Create a new custom metric.

fn profile_user_links_insert(&self, request: &EntityUserLink, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementProfileUserLinkInsertCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Adds a new user to the given view (profile).

fn webproperties_list(&self, account_id: &str) -> ManagementWebpropertyListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists web properties to which the user has access.

fn web_property_ad_words_links_insert(&self, request: &EntityAdWordsLink, account_id: &str, web_property_id: &str) -> ManagementWebPropertyAdWordsLinkInsertCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates a webProperty-AdWords link.

fn web_property_ad_words_links_patch(&self, request: &EntityAdWordsLink, account_id: &str, web_property_id: &str, web_property_ad_words_link_id: &str) -> ManagementWebPropertyAdWordsLinkPatchCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an existing webProperty-AdWords link. This method supports patch semantics.

fn goals_get(&self, account_id: &str, web_property_id: &str, profile_id: &str, goal_id: &str) -> ManagementGoalGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets a goal to which the user has access.

fn web_property_ad_words_links_delete(&self, account_id: &str, web_property_id: &str, web_property_ad_words_link_id: &str) -> ManagementWebPropertyAdWordsLinkDeleteCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes a web property-AdWords link.

fn web_property_ad_words_links_update(&self, request: &EntityAdWordsLink, account_id: &str, web_property_id: &str, web_property_ad_words_link_id: &str) -> ManagementWebPropertyAdWordsLinkUpdateCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an existing webProperty-AdWords link.

fn filters_patch(&self, request: &Filter, account_id: &str, filter_id: &str) -> ManagementFilterPatchCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an existing filter. This method supports patch semantics.

fn filters_list(&self, account_id: &str) -> ManagementFilterListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists all filters for an account

fn filters_delete(&self, account_id: &str, filter_id: &str) -> ManagementFilterDeleteCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Delete a filter.

fn profiles_insert(&self, request: &Profile, account_id: &str, web_property_id: &str) -> ManagementProfileInsertCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Create a new view (profile).

fn account_user_links_delete(&self, account_id: &str, link_id: &str) -> ManagementAccountUserLinkDeleteCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Removes a user from the given account.

fn profiles_update(&self, request: &Profile, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementProfileUpdateCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an existing view (profile).

fn profile_filter_links_get(&self, account_id: &str, web_property_id: &str, profile_id: &str, link_id: &str) -> ManagementProfileFilterLinkGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns a single profile filter link.