Struct google_analytics3::Profile

pub struct Profile {
    pub updated: Option<String>,
    pub exclude_query_parameters: Option<String>,
    pub currency: Option<String>,
    pub internal_web_property_id: Option<String>,
    pub child_link: Option<ProfileChildLink>,
    pub enhanced_e_commerce_tracking: Option<bool>,
    pub e_commerce_tracking: Option<bool>,
    pub web_property_id: Option<String>,
    pub timezone: Option<String>,
    pub id: Option<String>,
    pub account_id: Option<String>,
    pub default_page: Option<String>,
    pub kind: Option<String>,
    pub strip_site_search_query_parameters: Option<bool>,
    pub name: Option<String>,
    pub created: Option<String>,
    pub site_search_query_parameters: Option<String>,
    pub website_url: Option<String>,
    pub strip_site_search_category_parameters: Option<bool>,
    pub site_search_category_parameters: Option<String>,
    pub parent_link: Option<ProfileParentLink>,
    pub permissions: Option<ProfilePermissions>,
    pub type_: Option<String>,
    pub self_link: Option<String>,
}

JSON template for an Analytics view (profile).

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

profiles insert management (request|response)

profiles update management (request|response)

profiles get management (response)

profiles patch management (request|response)

Fields

updated

Time this view (profile) was last modified.

exclude_query_parameters

The query parameters that are excluded from this view (profile).

currency

The currency type associated with this view (profile). The supported values are: ARS, AUD, BGN, BRL, CAD, CHF, CNY, CZK, DKK, EUR, GBP, HKD, HUF, IDR, INR, JPY, KRW, LTL, MXN, NOK, NZD, PHP, PLN, RUB, SEK, THB, TRY, TWD, USD, VND, ZAR

internal_web_property_id

Internal ID for the web property to which this view (profile) belongs.

child_link

Child link for this view (profile). Points to the list of goals for this view (profile).

enhanced_e_commerce_tracking

Indicates whether enhanced ecommerce tracking is enabled for this view (profile). This property can only be enabled if ecommerce tracking is enabled.

e_commerce_tracking

Indicates whether ecommerce tracking is enabled for this view (profile).

web_property_id

Web property ID of the form UA-XXXXX-YY to which this view (profile) belongs.

timezone

Time zone for which this view (profile) has been configured. Time zones are identified by strings from the TZ database.

id

View (Profile) ID.

account_id

Account ID to which this view (profile) belongs.

default_page

Default page for this view (profile).

kind

Resource type for Analytics view (profile).

strip_site_search_query_parameters

Whether or not Analytics will strip search query parameters from the URLs in your reports.

name

Name of this view (profile).

created

Time this view (profile) was created.

site_search_query_parameters

The site search query parameters for this view (profile).

website_url

Website URL for this view (profile).

strip_site_search_category_parameters

Whether or not Analytics will strip search category parameters from the URLs in your reports.

site_search_category_parameters

Site search category parameters for this view (profile).

parent_link

Parent link for this view (profile). Points to the web property to which this view (profile) belongs.

permissions

Permissions the user has for this view (profile).

type_

View (Profile) type. Supported types: WEB or APP.

self_link

Link for this view (profile).