Struct google_analytics3::UnsampledReport

pub struct UnsampledReport {
    pub status: Option<String>,
    pub updated: Option<String>,
    pub download_type: Option<String>,
    pub start_date: Option<String>,
    pub drive_download_details: Option<UnsampledReportDriveDownloadDetails>,
    pub metrics: Option<String>,
    pub filters: Option<String>,
    pub web_property_id: Option<String>,
    pub segment: Option<String>,
    pub id: Option<String>,
    pub account_id: Option<String>,
    pub kind: Option<String>,
    pub cloud_storage_download_details: Option<UnsampledReportCloudStorageDownloadDetails>,
    pub dimensions: Option<String>,
    pub created: Option<String>,
    pub title: Option<String>,
    pub profile_id: Option<String>,
    pub end_date: Option<String>,
    pub self_link: Option<String>,
}

JSON template for Analytics unsampled report resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

unsampled reports insert management (request|response)

unsampled reports get management (response)

Fields

status

Status of this unsampled report. Possible values are PENDING, COMPLETED, or FAILED.

updated

Time this unsampled report was last modified.

download_type

The type of download you need to use for the report data file.

start_date

The start date for the unsampled report.

drive_download_details

Download details for a file stored in Google Drive.

metrics

The metrics for the unsampled report.

filters

The filters for the unsampled report.

web_property_id

Web property ID to which this unsampled report belongs. The web property ID is of the form UA-XXXXX-YY.

segment

The segment for the unsampled report.

id

Unsampled report ID.

account_id

Account ID to which this unsampled report belongs.

kind

Resource type for an Analytics unsampled report.

cloud_storage_download_details

Download details for a file stored in Google Cloud Storage.

dimensions

The dimensions for the unsampled report.

created

Time this unsampled report was created.

title

Title of the unsampled report.

profile_id

View (Profile) ID to which this unsampled report belongs.

end_date

The end date for the unsampled report.

self_link

Link for this unsampled report.