Struct google_analytics3::UnsampledReports

pub struct UnsampledReports {
    pub username: Option<String>,
    pub kind: Option<String>,
    pub items: Option<Vec<UnsampledReport>>,
    pub items_per_page: Option<i32>,
    pub previous_link: Option<String>,
    pub start_index: Option<i32>,
    pub next_link: Option<String>,
    pub total_results: Option<i32>,
}

An unsampled report collection lists Analytics unsampled reports to which the user has access. Each view (profile) can have a set of unsampled reports. Each resource in the unsampled report collection corresponds to a single Analytics unsampled report.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

unsampled reports list management (response)

Fields

username

Email ID of the authenticated user

kind

Collection type.

items

A list of unsampled reports.

items_per_page

The maximum number of resources the response can contain, regardless of the actual number of resources returned. Its value ranges from 1 to 1000 with a value of 1000 by default, or otherwise specified by the max-results query parameter.

previous_link

Link to previous page for this unsampled report collection.

start_index

The starting index of the resources, which is 1 by default or otherwise specified by the start-index query parameter.

next_link

Link to next page for this unsampled report collection.

total_results

The total number of results for the query, regardless of the number of resources in the result.