Crate google_analytics3

This documentation was generated from analytics crate version 2.0.0+20190807, where 20190807 is the exact revision of the analytics:v3 schema built by the mako code generator v2.0.0.

Everything else about the analytics v3 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

data

ga get, mcf get and realtime get

management

account summaries list, account user links delete, account user links insert, account user links list, account user links update, accounts list, client id hash client id, custom data sources list, custom dimensions get, custom dimensions insert, custom dimensions list, custom dimensions patch, custom dimensions update, custom metrics get, custom metrics insert, custom metrics list, custom metrics patch, custom metrics update, experiments delete, experiments get, experiments insert, experiments list, experiments patch, experiments update, filters delete, filters get, filters insert, filters list, filters patch, filters update, goals get, goals insert, goals list, goals patch, goals update, profile filter links delete, profile filter links get, profile filter links insert, profile filter links list, profile filter links patch, profile filter links update, profile user links delete, profile user links insert, profile user links list, profile user links update, profiles delete, profiles get, profiles insert, profiles list, profiles patch, profiles update, remarketing audience delete, remarketing audience get, remarketing audience insert, remarketing audience list, remarketing audience patch, remarketing audience update, segments list, unsampled reports delete, unsampled reports get, unsampled reports insert, unsampled reports list, uploads delete upload data, uploads get, uploads list, uploads upload data, web property ad words links delete, web property ad words links get, web property ad words links insert, web property ad words links list, web property ad words links patch, web property ad words links update, webproperties get, webproperties insert, webproperties list, webproperties patch, webproperties update, webproperty user links delete, webproperty user links insert, webproperty user links list and webproperty user links update

metadata

columns list

provisioning

create account ticket and create account tree

user deletion

user deletion request upsert

Upload supported by ...

uploads upload data management

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit().await

Or specifically ...

let r = hub.management().account_user_links_insert(...).doit().await
let r = hub.management().account_user_links_update(...).doit().await
let r = hub.management().profile_user_links_insert(...).doit().await
let r = hub.management().profile_user_links_update(...).doit().await
let r = hub.management().webproperty_user_links_insert(...).doit().await
let r = hub.management().webproperty_user_links_update(...).doit().await

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-analytics3 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.14"
hyper-rustls = "^0.22"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^5.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_analytics3 as analytics3;
use analytics3::api::EntityUserLink;
use analytics3::{Result, Error};
use std::default::Default;
use oauth2;
use analytics3::Analytics;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: oauth2::ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = Analytics::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = EntityUserLink::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.management().profile_user_links_update(req, "accountId", "webPropertyId", "profileId", "linkId")
             .doit().await;
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::Io(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be encodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Re-exports

pub use api::Analytics;

pub use client::Result;

pub use client::Error;

pub use client::Delegate;

Modules

api

client