Struct google_analytics3::CustomDataSource

pub struct CustomDataSource {
    pub kind: Option<String>,
    pub description: Option<String>,
    pub profiles_linked: Option<Vec<String>>,
    pub upload_type: Option<String>,
    pub created: Option<String>,
    pub updated: Option<String>,
    pub account_id: Option<String>,
    pub child_link: Option<CustomDataSourceChildLink>,
    pub web_property_id: Option<String>,
    pub parent_link: Option<CustomDataSourceParentLink>,
    pub import_behavior: Option<String>,
    pub type_: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub name: Option<String>,
}

JSON template for an Analytics custom data source.

This type is not used in any activity, and only used as part of another schema.

Fields
kind: Option<String>

Resource type for Analytics custom data source.

description: Option<String>

Description of custom data source.

profiles_linked: Option<Vec<String>>

IDs of views (profiles) linked to the custom data source.

upload_type: Option<String>

no description provided

created: Option<String>

Time this custom data source was created.

updated: Option<String>

Time this custom data source was last modified.

account_id: Option<String>

Account ID to which this custom data source belongs.

child_link: Option<CustomDataSourceChildLink>

no description provided

web_property_id: Option<String>

Web property ID of the form UA-XXXXX-YY to which this custom data source belongs.

parent_link: Option<CustomDataSourceParentLink>

Parent link for this custom data source. Points to the web property to which this custom data source belongs.

import_behavior: Option<String>

no description provided

type_: Option<String>

Type of the custom data source.

id: Option<String>

Custom data source ID.

self_link: Option<String>

Link for this Analytics custom data source.

name: Option<String>

Name of this custom data source.