Struct google_analytics3::Experiment

pub struct Experiment {
    pub status: Option<String>,
    pub traffic_coverage: Option<f64>,
    pub description: Option<String>,
    pub web_property_id: Option<String>,
    pub updated: Option<String>,
    pub internal_web_property_id: Option<String>,
    pub winner_confidence_level: Option<f64>,
    pub start_time: Option<String>,
    pub optimization_type: Option<String>,
    pub winner_found: Option<bool>,
    pub objective_metric: Option<String>,
    pub id: Option<String>,
    pub equal_weighting: Option<bool>,
    pub account_id: Option<String>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub created: Option<String>,
    pub reason_experiment_ended: Option<String>,
    pub variations: Option<Vec<ExperimentVariations>>,
    pub snippet: Option<String>,
    pub editable_in_ga_ui: Option<bool>,
    pub rewrite_variation_urls_as_original: Option<bool>,
    pub minimum_experiment_length_in_days: Option<i32>,
    pub profile_id: Option<String>,
    pub parent_link: Option<ExperimentParentLink>,
    pub end_time: Option<String>,
    pub serving_framework: Option<String>,
    pub self_link: Option<String>,
}

JSON template for Analytics experiment resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

experiments patch management (request|response)

experiments insert management (request|response)

experiments update management (request|response)

experiments get management (response)

Fields

status

Experiment status. Possible values: "DRAFT", "READY_TO_RUN", "RUNNING", "ENDED". Experiments can be created in the "DRAFT", "READY_TO_RUN" or "RUNNING" state. This field is required when creating an experiment.

traffic_coverage

A floating-point number in (0, 1]. Specifies the fraction of the traffic that participates in the experiment. Can be changed for a running experiment. This field may not be changed for an experiments whose status is ENDED.

description

Notes about this experiment.

web_property_id

Web property ID to which this experiment belongs. The web property ID is of the form UA-XXXXX-YY. This field is read-only.

updated

Time the experiment was last modified. This field is read-only.

internal_web_property_id

Internal ID for the web property to which this experiment belongs. This field is read-only.

winner_confidence_level

A floating-point number in (0, 1). Specifies the necessary confidence level to choose a winner. This field may not be changed for an experiments whose status is ENDED.

start_time

The starting time of the experiment (the time the status changed from READY_TO_RUN to RUNNING). This field is present only if the experiment has started. This field is read-only.

optimization_type

Whether the objectiveMetric should be minimized or maximized. Possible values: "MAXIMUM", "MINIMUM". Optional--defaults to "MAXIMUM". Cannot be specified without objectiveMetric. Cannot be modified when status is "RUNNING" or "ENDED".

winner_found

Boolean specifying whether a winner has been found for this experiment. This field is read-only.

objective_metric

The metric that the experiment is optimizing. Valid values: "ga:goal(n)Completions", "ga:adsenseAdsClicks", "ga:adsenseAdsViewed", "ga:adsenseRevenue", "ga:bounces", "ga:pageviews", "ga:sessionDuration", "ga:transactions", "ga:transactionRevenue". This field is required if status is "RUNNING" and servingFramework is one of "REDIRECT" or "API".

id

Experiment ID. Required for patch and update. Disallowed for create.

equal_weighting

Boolean specifying whether to distribute traffic evenly across all variations. If the value is False, content experiments follows the default behavior of adjusting traffic dynamically based on variation performance. Optional -- defaults to False. This field may not be changed for an experiment whose status is ENDED.

account_id

Account ID to which this experiment belongs. This field is read-only.

kind

Resource type for an Analytics experiment. This field is read-only.

name

Experiment name. This field may not be changed for an experiment whose status is ENDED. This field is required when creating an experiment.

created

Time the experiment was created. This field is read-only.

reason_experiment_ended

Why the experiment ended. Possible values: "STOPPED_BY_USER", "WINNER_FOUND", "EXPERIMENT_EXPIRED", "ENDED_WITH_NO_WINNER", "GOAL_OBJECTIVE_CHANGED". "ENDED_WITH_NO_WINNER" means that the experiment didn't expire but no winner was projected to be found. If the experiment status is changed via the API to ENDED this field is set to STOPPED_BY_USER. This field is read-only.

variations

Array of variations. The first variation in the array is the original. The number of variations may not change once an experiment is in the RUNNING state. At least two variations are required before status can be set to RUNNING.

snippet

The snippet of code to include on the control page(s). This field is read-only.

editable_in_ga_ui

If true, the end user will be able to edit the experiment via the Google Analytics user interface.

rewrite_variation_urls_as_original

Boolean specifying whether variations URLS are rewritten to match those of the original. This field may not be changed for an experiments whose status is ENDED.

minimum_experiment_length_in_days

An integer number in [3, 90]. Specifies the minimum length of the experiment. Can be changed for a running experiment. This field may not be changed for an experiments whose status is ENDED.

profile_id

View (Profile) ID to which this experiment belongs. This field is read-only.

parent_link

Parent link for an experiment. Points to the view (profile) to which this experiment belongs.

end_time

The ending time of the experiment (the time the status changed from RUNNING to ENDED). This field is present only if the experiment has ended. This field is read-only.

serving_framework

The framework used to serve the experiment variations and evaluate the results. One of: - REDIRECT: Google Analytics redirects traffic to different variation pages, reports the chosen variation and evaluates the results. - API: Google Analytics chooses and reports the variation to serve and evaluates the results; the caller is responsible for serving the selected variation. - EXTERNAL: The variations will be served externally and the chosen variation reported to Google Analytics. The caller is responsible for serving the selected variation and evaluating the results.

self_link

Link for this experiment. This field is read-only.