Struct google_analytics3::ExperimentVariations

pub struct ExperimentVariations {
    pub status: Option<String>,
    pub url: Option<String>,
    pub won: Option<bool>,
    pub name: Option<String>,
    pub weight: Option<f64>,
}

Array of variations. The first variation in the array is the original. The number of variations may not change once an experiment is in the RUNNING state. At least two variations are required before status can be set to RUNNING.

This type is not used in any activity, and only used as part of another schema.

Fields

status

Status of the variation. Possible values: "ACTIVE", "INACTIVE". INACTIVE variations are not served. This field may not be changed for an experiment whose status is ENDED.

url

The URL of the variation. This field may not be changed for an experiment whose status is RUNNING or ENDED.

won

True if the experiment has ended and this variation performed (statistically) significantly better than the original. This field is read-only.

name

The name of the variation. This field is required when creating an experiment. This field may not be changed for an experiment whose status is ENDED.

weight

Weight that this variation should receive. Only present if the experiment is running. This field is read-only.