Struct google_analytics3::ManagementMethods

pub struct ManagementMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on management resources. It is not used directly, but through the Analytics hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_analytics3 as analytics3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use analytics3::Analytics;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Analytics::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `account_summaries_list(...)`, `account_user_links_delete(...)`, `account_user_links_insert(...)`, `account_user_links_list(...)`, `account_user_links_update(...)`, `accounts_list(...)`, `custom_data_sources_list(...)`, `custom_dimensions_get(...)`, `custom_dimensions_insert(...)`, `custom_dimensions_list(...)`, `custom_dimensions_patch(...)`, `custom_dimensions_update(...)`, `custom_metrics_get(...)`, `custom_metrics_insert(...)`, `custom_metrics_list(...)`, `custom_metrics_patch(...)`, `custom_metrics_update(...)`, `experiments_delete(...)`, `experiments_get(...)`, `experiments_insert(...)`, `experiments_list(...)`, `experiments_patch(...)`, `experiments_update(...)`, `filters_delete(...)`, `filters_get(...)`, `filters_insert(...)`, `filters_list(...)`, `filters_patch(...)`, `filters_update(...)`, `goals_get(...)`, `goals_insert(...)`, `goals_list(...)`, `goals_patch(...)`, `goals_update(...)`, `profile_filter_links_delete(...)`, `profile_filter_links_get(...)`, `profile_filter_links_insert(...)`, `profile_filter_links_list(...)`, `profile_filter_links_patch(...)`, `profile_filter_links_update(...)`, `profile_user_links_delete(...)`, `profile_user_links_insert(...)`, `profile_user_links_list(...)`, `profile_user_links_update(...)`, `profiles_delete(...)`, `profiles_get(...)`, `profiles_insert(...)`, `profiles_list(...)`, `profiles_patch(...)`, `profiles_update(...)`, `segments_list(...)`, `unsampled_reports_delete(...)`, `unsampled_reports_get(...)`, `unsampled_reports_insert(...)`, `unsampled_reports_list(...)`, `uploads_delete_upload_data(...)`, `uploads_get(...)`, `uploads_list(...)`, `uploads_upload_data(...)`, `web_property_ad_words_links_delete(...)`, `web_property_ad_words_links_get(...)`, `web_property_ad_words_links_insert(...)`, `web_property_ad_words_links_list(...)`, `web_property_ad_words_links_patch(...)`, `web_property_ad_words_links_update(...)`, `webproperties_get(...)`, `webproperties_insert(...)`, `webproperties_list(...)`, `webproperties_patch(...)`, `webproperties_update(...)`, `webproperty_user_links_delete(...)`, `webproperty_user_links_insert(...)`, `webproperty_user_links_list(...)` and `webproperty_user_links_update(...)`
// to build up your call.
let rb = hub.management();

Methods

impl<'a, C, A> ManagementMethods<'a, C, A>

fn webproperties_insert(&self, request: Webproperty, account_id: &str) -> ManagementWebpropertyInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Create a new property if the account has fewer than 20 properties. Web properties are visible in the Google Analytics interface only if they have at least one profile.

Arguments

request - No description provided.

accountId - Account ID to create the web property for.

fn profiles_get(&self, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementProfileGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a view (profile) to which the user has access.

Arguments

accountId - Account ID to retrieve the view (profile) for.

webPropertyId - Web property ID to retrieve the view (profile) for.

profileId - View (Profile) ID to retrieve the view (profile) for.

fn accounts_list(&self) -> ManagementAccountListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all accounts to which the user has access.

fn profile_filter_links_patch(&self, request: ProfileFilterLink, account_id: &str, web_property_id: &str, profile_id: &str, link_id: &str) -> ManagementProfileFilterLinkPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Update an existing profile filter link. This method supports patch semantics.

Arguments

request - No description provided.

accountId - Account ID to which profile filter link belongs.

webPropertyId - Web property Id to which profile filter link belongs

profileId - Profile ID to which filter link belongs

linkId - ID of the profile filter link to be updated.

fn webproperty_user_links_delete(&self, account_id: &str, web_property_id: &str, link_id: &str) -> ManagementWebpropertyUserLinkDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Removes a user from the given web property.

Arguments

accountId - Account ID to delete the user link for.

webPropertyId - Web Property ID to delete the user link for.

linkId - Link ID to delete the user link for.

fn profile_user_links_delete(&self, account_id: &str, web_property_id: &str, profile_id: &str, link_id: &str) -> ManagementProfileUserLinkDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Removes a user from the given view (profile).

Arguments

accountId - Account ID to delete the user link for.

webPropertyId - Web Property ID to delete the user link for.

profileId - View (Profile) ID to delete the user link for.

linkId - Link ID to delete the user link for.

fn profile_user_links_update(&self, request: EntityUserLink, account_id: &str, web_property_id: &str, profile_id: &str, link_id: &str) -> ManagementProfileUserLinkUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates permissions for an existing user on the given view (profile).

Arguments

request - No description provided.

accountId - Account ID to update the user link for.

webPropertyId - Web Property ID to update the user link for.

profileId - View (Profile ID) to update the user link for.

linkId - Link ID to update the user link for.

fn filters_insert(&self, request: Filter, account_id: &str) -> ManagementFilterInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Create a new filter.

Arguments

request - No description provided.

accountId - Account ID to create filter for.

fn account_user_links_update(&self, request: EntityUserLink, account_id: &str, link_id: &str) -> ManagementAccountUserLinkUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates permissions for an existing user on the given account.

Arguments

request - No description provided.

accountId - Account ID to update the account-user link for.

linkId - Link ID to update the account-user link for.

fn webproperties_update(&self, request: Webproperty, account_id: &str, web_property_id: &str) -> ManagementWebpropertyUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing web property.

Arguments

request - No description provided.

accountId - Account ID to which the web property belongs

webPropertyId - Web property ID

fn webproperty_user_links_update(&self, request: EntityUserLink, account_id: &str, web_property_id: &str, link_id: &str) -> ManagementWebpropertyUserLinkUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates permissions for an existing user on the given web property.

Arguments

request - No description provided.

accountId - Account ID to update the account-user link for.

webPropertyId - Web property ID to update the account-user link for.

linkId - Link ID to update the account-user link for.

fn unsampled_reports_insert(&self, request: UnsampledReport, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementUnsampledReportInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Create a new unsampled report.

Arguments

request - No description provided.

accountId - Account ID to create the unsampled report for.

webPropertyId - Web property ID to create the unsampled report for.

profileId - View (Profile) ID to create the unsampled report for.

fn custom_metrics_get(&self, account_id: &str, web_property_id: &str, custom_metric_id: &str) -> ManagementCustomMetricGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get a custom metric to which the user has access.

Arguments

accountId - Account ID for the custom metric to retrieve.

webPropertyId - Web property ID for the custom metric to retrieve.

customMetricId - The ID of the custom metric to retrieve.

fn uploads_get(&self, account_id: &str, web_property_id: &str, custom_data_source_id: &str, upload_id: &str) -> ManagementUploadGetCall<'a, C, A>

Create a builder to help you perform the following task:

List uploads to which the user has access.

Arguments

accountId - Account Id for the upload to retrieve.

webPropertyId - Web property Id for the upload to retrieve.

customDataSourceId - Custom data source Id for upload to retrieve.

uploadId - Upload Id to retrieve.

fn web_property_ad_words_links_get(&self, account_id: &str, web_property_id: &str, web_property_ad_words_link_id: &str) -> ManagementWebPropertyAdWordsLinkGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a web property-AdWords link to which the user has access.

Arguments

accountId - ID of the account which the given web property belongs to.

webPropertyId - Web property ID to retrieve the AdWords link for.

webPropertyAdWordsLinkId - Web property-AdWords link ID.

fn webproperty_user_links_list(&self, account_id: &str, web_property_id: &str) -> ManagementWebpropertyUserLinkListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists webProperty-user links for a given web property.

Arguments

accountId - Account ID which the given web property belongs to.

webPropertyId - Web Property ID for the webProperty-user links to retrieve. Can either be a specific web property ID or '~all', which refers to all the web properties that user has access to.

fn profile_filter_links_update(&self, request: ProfileFilterLink, account_id: &str, web_property_id: &str, profile_id: &str, link_id: &str) -> ManagementProfileFilterLinkUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Update an existing profile filter link.

Arguments

request - No description provided.

accountId - Account ID to which profile filter link belongs.

webPropertyId - Web property Id to which profile filter link belongs

profileId - Profile ID to which filter link belongs

linkId - ID of the profile filter link to be updated.

fn experiments_get(&self, account_id: &str, web_property_id: &str, profile_id: &str, experiment_id: &str) -> ManagementExperimentGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns an experiment to which the user has access.

Arguments

accountId - Account ID to retrieve the experiment for.

webPropertyId - Web property ID to retrieve the experiment for.

profileId - View (Profile) ID to retrieve the experiment for.

experimentId - Experiment ID to retrieve the experiment for.

fn custom_dimensions_update(&self, request: CustomDimension, account_id: &str, web_property_id: &str, custom_dimension_id: &str) -> ManagementCustomDimensionUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing custom dimension.

Arguments

request - No description provided.

accountId - Account ID for the custom dimension to update.

webPropertyId - Web property ID for the custom dimension to update.

customDimensionId - Custom dimension ID for the custom dimension to update.

fn unsampled_reports_get(&self, account_id: &str, web_property_id: &str, profile_id: &str, unsampled_report_id: &str) -> ManagementUnsampledReportGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a single unsampled report.

Arguments

accountId - Account ID to retrieve unsampled report for.

webPropertyId - Web property ID to retrieve unsampled reports for.

profileId - View (Profile) ID to retrieve unsampled report for.

unsampledReportId - ID of the unsampled report to retrieve.

fn profile_filter_links_insert(&self, request: ProfileFilterLink, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementProfileFilterLinkInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Create a new profile filter link.

Arguments

request - No description provided.

accountId - Account ID to create profile filter link for.

webPropertyId - Web property Id to create profile filter link for.

profileId - Profile ID to create filter link for.

fn filters_update(&self, request: Filter, account_id: &str, filter_id: &str) -> ManagementFilterUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing filter.

Arguments

request - No description provided.

accountId - Account ID to which the filter belongs.

filterId - ID of the filter to be updated.

fn profile_filter_links_delete(&self, account_id: &str, web_property_id: &str, profile_id: &str, link_id: &str) -> ManagementProfileFilterLinkDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete a profile filter link.

Arguments

accountId - Account ID to which the profile filter link belongs.

webPropertyId - Web property Id to which the profile filter link belongs.

profileId - Profile ID to which the filter link belongs.

linkId - ID of the profile filter link to delete.

fn segments_list(&self) -> ManagementSegmentListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists segments to which the user has access.

fn profiles_delete(&self, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementProfileDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a view (profile).

Arguments

accountId - Account ID to delete the view (profile) for.

webPropertyId - Web property ID to delete the view (profile) for.

profileId - ID of the view (profile) to be deleted.

fn goals_patch(&self, request: Goal, account_id: &str, web_property_id: &str, profile_id: &str, goal_id: &str) -> ManagementGoalPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing goal. This method supports patch semantics.

Arguments

request - No description provided.

accountId - Account ID to update the goal.

webPropertyId - Web property ID to update the goal.

profileId - View (Profile) ID to update the goal.

goalId - Index of the goal to be updated.

fn custom_dimensions_patch(&self, request: CustomDimension, account_id: &str, web_property_id: &str, custom_dimension_id: &str) -> ManagementCustomDimensionPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing custom dimension. This method supports patch semantics.

Arguments

request - No description provided.

accountId - Account ID for the custom dimension to update.

webPropertyId - Web property ID for the custom dimension to update.

customDimensionId - Custom dimension ID for the custom dimension to update.

fn experiments_patch(&self, request: Experiment, account_id: &str, web_property_id: &str, profile_id: &str, experiment_id: &str) -> ManagementExperimentPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Update an existing experiment. This method supports patch semantics.

Arguments

request - No description provided.

accountId - Account ID of the experiment to update.

webPropertyId - Web property ID of the experiment to update.

profileId - View (Profile) ID of the experiment to update.

experimentId - Experiment ID of the experiment to update.

fn experiments_list(&self, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementExperimentListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists experiments to which the user has access.

Arguments

accountId - Account ID to retrieve experiments for.

webPropertyId - Web property ID to retrieve experiments for.

profileId - View (Profile) ID to retrieve experiments for.

fn profiles_list(&self, account_id: &str, web_property_id: &str) -> ManagementProfileListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists views (profiles) to which the user has access.

Arguments

accountId - Account ID for the view (profiles) to retrieve. Can either be a specific account ID or '~all', which refers to all the accounts to which the user has access.

webPropertyId - Web property ID for the views (profiles) to retrieve. Can either be a specific web property ID or '~all', which refers to all the web properties to which the user has access.

fn goals_list(&self, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementGoalListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists goals to which the user has access.

Arguments

accountId - Account ID to retrieve goals for. Can either be a specific account ID or '~all', which refers to all the accounts that user has access to.

webPropertyId - Web property ID to retrieve goals for. Can either be a specific web property ID or '~all', which refers to all the web properties that user has access to.

profileId - View (Profile) ID to retrieve goals for. Can either be a specific view (profile) ID or '~all', which refers to all the views (profiles) that user has access to.

fn goals_insert(&self, request: Goal, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementGoalInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Create a new goal.

Arguments

request - No description provided.

accountId - Account ID to create the goal for.

webPropertyId - Web property ID to create the goal for.

profileId - View (Profile) ID to create the goal for.

fn custom_metrics_patch(&self, request: CustomMetric, account_id: &str, web_property_id: &str, custom_metric_id: &str) -> ManagementCustomMetricPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing custom metric. This method supports patch semantics.

Arguments

request - No description provided.

accountId - Account ID for the custom metric to update.

webPropertyId - Web property ID for the custom metric to update.

customMetricId - Custom metric ID for the custom metric to update.

fn uploads_upload_data(&self, account_id: &str, web_property_id: &str, custom_data_source_id: &str) -> ManagementUploadUploadDataCall<'a, C, A>

Create a builder to help you perform the following task:

Upload data for a custom data source.

Arguments

accountId - Account Id associated with the upload.

webPropertyId - Web property UA-string associated with the upload.

customDataSourceId - Custom data source Id to which the data being uploaded belongs.

fn account_user_links_insert(&self, request: EntityUserLink, account_id: &str) -> ManagementAccountUserLinkInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Adds a new user to the given account.

Arguments

request - No description provided.

accountId - Account ID to create the user link for.

fn webproperty_user_links_insert(&self, request: EntityUserLink, account_id: &str, web_property_id: &str) -> ManagementWebpropertyUserLinkInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Adds a new user to the given web property.

Arguments

request - No description provided.

accountId - Account ID to create the user link for.

webPropertyId - Web Property ID to create the user link for.

fn experiments_delete(&self, account_id: &str, web_property_id: &str, profile_id: &str, experiment_id: &str) -> ManagementExperimentDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete an experiment.

Arguments

accountId - Account ID to which the experiment belongs

webPropertyId - Web property ID to which the experiment belongs

profileId - View (Profile) ID to which the experiment belongs

experimentId - ID of the experiment to delete

fn unsampled_reports_delete(&self, account_id: &str, web_property_id: &str, profile_id: &str, unsampled_report_id: &str) -> ManagementUnsampledReportDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an unsampled report.

Arguments

accountId - Account ID to delete the unsampled report for.

webPropertyId - Web property ID to delete the unsampled reports for.

profileId - View (Profile) ID to delete the unsampled report for.

unsampledReportId - ID of the unsampled report to be deleted.

fn profiles_patch(&self, request: Profile, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementProfilePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing view (profile). This method supports patch semantics.

Arguments

request - No description provided.

accountId - Account ID to which the view (profile) belongs

webPropertyId - Web property ID to which the view (profile) belongs

profileId - ID of the view (profile) to be updated.

fn filters_get(&self, account_id: &str, filter_id: &str) -> ManagementFilterGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a filters to which the user has access.

Arguments

accountId - Account ID to retrieve filters for.

filterId - Filter ID to retrieve filters for.

fn web_property_ad_words_links_list(&self, account_id: &str, web_property_id: &str) -> ManagementWebPropertyAdWordsLinkListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists webProperty-AdWords links for a given web property.

Arguments

accountId - ID of the account which the given web property belongs to.

webPropertyId - Web property ID to retrieve the AdWords links for.

fn account_summaries_list(&self) -> ManagementAccountSummaryListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists account summaries (lightweight tree comprised of accounts/properties/profiles) to which the user has access.

fn custom_dimensions_list(&self, account_id: &str, web_property_id: &str) -> ManagementCustomDimensionListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists custom dimensions to which the user has access.

Arguments

accountId - Account ID for the custom dimensions to retrieve.

webPropertyId - Web property ID for the custom dimensions to retrieve.

fn uploads_list(&self, account_id: &str, web_property_id: &str, custom_data_source_id: &str) -> ManagementUploadListCall<'a, C, A>

Create a builder to help you perform the following task:

List uploads to which the user has access.

Arguments

accountId - Account Id for the uploads to retrieve.

webPropertyId - Web property Id for the uploads to retrieve.

customDataSourceId - Custom data source Id for uploads to retrieve.

fn profile_user_links_list(&self, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementProfileUserLinkListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists profile-user links for a given view (profile).

Arguments

accountId - Account ID which the given view (profile) belongs to.

webPropertyId - Web Property ID which the given view (profile) belongs to. Can either be a specific web property ID or '~all', which refers to all the web properties that user has access to.

profileId - View (Profile) ID to retrieve the profile-user links for. Can either be a specific profile ID or '~all', which refers to all the profiles that user has access to.

fn account_user_links_list(&self, account_id: &str) -> ManagementAccountUserLinkListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists account-user links for a given account.

Arguments

accountId - Account ID to retrieve the user links for.

fn custom_metrics_update(&self, request: CustomMetric, account_id: &str, web_property_id: &str, custom_metric_id: &str) -> ManagementCustomMetricUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing custom metric.

Arguments

request - No description provided.

accountId - Account ID for the custom metric to update.

webPropertyId - Web property ID for the custom metric to update.

customMetricId - Custom metric ID for the custom metric to update.

fn uploads_delete_upload_data(&self, request: AnalyticsDataimportDeleteUploadDataRequest, account_id: &str, web_property_id: &str, custom_data_source_id: &str) -> ManagementUploadDeleteUploadDataCall<'a, C, A>

Create a builder to help you perform the following task:

Delete data associated with a previous upload.

Arguments

request - No description provided.

accountId - Account Id for the uploads to be deleted.

webPropertyId - Web property Id for the uploads to be deleted.

customDataSourceId - Custom data source Id for the uploads to be deleted.

fn experiments_update(&self, request: Experiment, account_id: &str, web_property_id: &str, profile_id: &str, experiment_id: &str) -> ManagementExperimentUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Update an existing experiment.

Arguments

request - No description provided.

accountId - Account ID of the experiment to update.

webPropertyId - Web property ID of the experiment to update.

profileId - View (Profile) ID of the experiment to update.

experimentId - Experiment ID of the experiment to update.

fn goals_update(&self, request: Goal, account_id: &str, web_property_id: &str, profile_id: &str, goal_id: &str) -> ManagementGoalUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing goal.

Arguments

request - No description provided.

accountId - Account ID to update the goal.

webPropertyId - Web property ID to update the goal.

profileId - View (Profile) ID to update the goal.

goalId - Index of the goal to be updated.

fn custom_data_sources_list(&self, account_id: &str, web_property_id: &str) -> ManagementCustomDataSourceListCall<'a, C, A>

Create a builder to help you perform the following task:

List custom data sources to which the user has access.

Arguments

accountId - Account Id for the custom data sources to retrieve.

webPropertyId - Web property Id for the custom data sources to retrieve.

fn custom_metrics_list(&self, account_id: &str, web_property_id: &str) -> ManagementCustomMetricListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists custom metrics to which the user has access.

Arguments

accountId - Account ID for the custom metrics to retrieve.

webPropertyId - Web property ID for the custom metrics to retrieve.

fn unsampled_reports_list(&self, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementUnsampledReportListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists unsampled reports to which the user has access.

Arguments

accountId - Account ID to retrieve unsampled reports for. Must be a specific account ID, ~all is not supported.

webPropertyId - Web property ID to retrieve unsampled reports for. Must be a specific web property ID, ~all is not supported.

profileId - View (Profile) ID to retrieve unsampled reports for. Must be a specific view (profile) ID, ~all is not supported.

fn webproperties_get(&self, account_id: &str, web_property_id: &str) -> ManagementWebpropertyGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a web property to which the user has access.

Arguments

accountId - Account ID to retrieve the web property for.

webPropertyId - ID to retrieve the web property for.

fn experiments_insert(&self, request: Experiment, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementExperimentInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Create a new experiment.

Arguments

request - No description provided.

accountId - Account ID to create the experiment for.

webPropertyId - Web property ID to create the experiment for.

profileId - View (Profile) ID to create the experiment for.

fn custom_dimensions_get(&self, account_id: &str, web_property_id: &str, custom_dimension_id: &str) -> ManagementCustomDimensionGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get a custom dimension to which the user has access.

Arguments

accountId - Account ID for the custom dimension to retrieve.

webPropertyId - Web property ID for the custom dimension to retrieve.

customDimensionId - The ID of the custom dimension to retrieve.

fn profile_filter_links_list(&self, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementProfileFilterLinkListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all profile filter links for a profile.

Arguments

accountId - Account ID to retrieve profile filter links for.

webPropertyId - Web property Id for profile filter links for. Can either be a specific web property ID or '~all', which refers to all the web properties that user has access to.

profileId - Profile ID to retrieve filter links for. Can either be a specific profile ID or '~all', which refers to all the profiles that user has access to.

fn webproperties_patch(&self, request: Webproperty, account_id: &str, web_property_id: &str) -> ManagementWebpropertyPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing web property. This method supports patch semantics.

Arguments

request - No description provided.

accountId - Account ID to which the web property belongs

webPropertyId - Web property ID

fn custom_dimensions_insert(&self, request: CustomDimension, account_id: &str, web_property_id: &str) -> ManagementCustomDimensionInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Create a new custom dimension.

Arguments

request - No description provided.

accountId - Account ID for the custom dimension to create.

webPropertyId - Web property ID for the custom dimension to create.

fn custom_metrics_insert(&self, request: CustomMetric, account_id: &str, web_property_id: &str) -> ManagementCustomMetricInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Create a new custom metric.

Arguments

request - No description provided.

accountId - Account ID for the custom metric to create.

webPropertyId - Web property ID for the custom dimension to create.

fn profile_user_links_insert(&self, request: EntityUserLink, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementProfileUserLinkInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Adds a new user to the given view (profile).

Arguments

request - No description provided.

accountId - Account ID to create the user link for.

webPropertyId - Web Property ID to create the user link for.

profileId - View (Profile) ID to create the user link for.

fn webproperties_list(&self, account_id: &str) -> ManagementWebpropertyListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists web properties to which the user has access.

Arguments

accountId - Account ID to retrieve web properties for. Can either be a specific account ID or '~all', which refers to all the accounts that user has access to.

fn web_property_ad_words_links_insert(&self, request: EntityAdWordsLink, account_id: &str, web_property_id: &str) -> ManagementWebPropertyAdWordsLinkInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a webProperty-AdWords link.

Arguments

request - No description provided.

accountId - ID of the Google Analytics account to create the link for.

webPropertyId - Web property ID to create the link for.

fn web_property_ad_words_links_patch(&self, request: EntityAdWordsLink, account_id: &str, web_property_id: &str, web_property_ad_words_link_id: &str) -> ManagementWebPropertyAdWordsLinkPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing webProperty-AdWords link. This method supports patch semantics.

Arguments

request - No description provided.

accountId - ID of the account which the given web property belongs to.

webPropertyId - Web property ID to retrieve the AdWords link for.

webPropertyAdWordsLinkId - Web property-AdWords link ID.

fn goals_get(&self, account_id: &str, web_property_id: &str, profile_id: &str, goal_id: &str) -> ManagementGoalGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a goal to which the user has access.

Arguments

accountId - Account ID to retrieve the goal for.

webPropertyId - Web property ID to retrieve the goal for.

profileId - View (Profile) ID to retrieve the goal for.

goalId - Goal ID to retrieve the goal for.

fn web_property_ad_words_links_delete(&self, account_id: &str, web_property_id: &str, web_property_ad_words_link_id: &str) -> ManagementWebPropertyAdWordsLinkDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a web property-AdWords link.

Arguments

accountId - ID of the account which the given web property belongs to.

webPropertyId - Web property ID to delete the AdWords link for.

webPropertyAdWordsLinkId - Web property AdWords link ID.

fn web_property_ad_words_links_update(&self, request: EntityAdWordsLink, account_id: &str, web_property_id: &str, web_property_ad_words_link_id: &str) -> ManagementWebPropertyAdWordsLinkUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing webProperty-AdWords link.

Arguments

request - No description provided.

accountId - ID of the account which the given web property belongs to.

webPropertyId - Web property ID to retrieve the AdWords link for.

webPropertyAdWordsLinkId - Web property-AdWords link ID.

fn filters_patch(&self, request: Filter, account_id: &str, filter_id: &str) -> ManagementFilterPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing filter. This method supports patch semantics.

Arguments

request - No description provided.

accountId - Account ID to which the filter belongs.

filterId - ID of the filter to be updated.

fn filters_list(&self, account_id: &str) -> ManagementFilterListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all filters for an account

Arguments

accountId - Account ID to retrieve filters for.

fn filters_delete(&self, account_id: &str, filter_id: &str) -> ManagementFilterDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete a filter.

Arguments

accountId - Account ID to delete the filter for.

filterId - ID of the filter to be deleted.

fn profiles_insert(&self, request: Profile, account_id: &str, web_property_id: &str) -> ManagementProfileInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Create a new view (profile).

Arguments

request - No description provided.

accountId - Account ID to create the view (profile) for.

webPropertyId - Web property ID to create the view (profile) for.

fn account_user_links_delete(&self, account_id: &str, link_id: &str) -> ManagementAccountUserLinkDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Removes a user from the given account.

Arguments

accountId - Account ID to delete the user link for.

linkId - Link ID to delete the user link for.

fn profiles_update(&self, request: Profile, account_id: &str, web_property_id: &str, profile_id: &str) -> ManagementProfileUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing view (profile).

Arguments

request - No description provided.

accountId - Account ID to which the view (profile) belongs

webPropertyId - Web property ID to which the view (profile) belongs

profileId - ID of the view (profile) to be updated.

fn profile_filter_links_get(&self, account_id: &str, web_property_id: &str, profile_id: &str, link_id: &str) -> ManagementProfileFilterLinkGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a single profile filter link.

Arguments

accountId - Account ID to retrieve profile filter link for.

webPropertyId - Web property Id to retrieve profile filter link for.

profileId - Profile ID to retrieve filter link for.

linkId - ID of the profile filter link.