Struct google_analytics3::RealtimeData

pub struct RealtimeData {
    pub kind: Option<String>,
    pub rows: Option<Vec<Vec<String>>>,
    pub total_results: Option<i32>,
    pub profile_info: Option<RealtimeDataProfileInfo>,
    pub column_headers: Option<Vec<RealtimeDataColumnHeaders>>,
    pub query: Option<RealtimeDataQuery>,
    pub totals_for_all_results: Option<HashMap<String, String>>,
    pub id: Option<String>,
    pub self_link: Option<String>,
}

Real time data for a given view (profile).

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

realtime get data (response)

Fields
kind: Option<String>

Resource type.

rows: Option<Vec<Vec<String>>>

Real time data rows, where each row contains a list of dimension values followed by the metric values. The order of dimensions and metrics is same as specified in the request.

total_results: Option<i32>

The total number of rows for the query, regardless of the number of rows in the response.

profile_info: Option<RealtimeDataProfileInfo>

Information for the view (profile), for which the real time data was requested.

column_headers: Option<Vec<RealtimeDataColumnHeaders>>

Column headers that list dimension names followed by the metric names. The order of dimensions and metrics is same as specified in the request.

query: Option<RealtimeDataQuery>

Real time data request query parameters.

totals_for_all_results: Option<HashMap<String, String>>

Total values for the requested metrics over all the results, not just the results returned in this response. The order of the metric totals is same as the metric order specified in the request.

id: Option<String>

Unique ID for this data response.

self_link: Option<String>

Link to this page.