Struct google_analytics3::UnsampledReport

pub struct UnsampledReport {
    pub status: Option<String>,
    pub updated: Option<String>,
    pub download_type: Option<String>,
    pub start_date: Option<String>,
    pub drive_download_details: Option<UnsampledReportDriveDownloadDetails>,
    pub metrics: Option<String>,
    pub filters: Option<String>,
    pub web_property_id: Option<String>,
    pub segment: Option<String>,
    pub id: Option<String>,
    pub account_id: Option<String>,
    pub kind: Option<String>,
    pub cloud_storage_download_details: Option<UnsampledReportCloudStorageDownloadDetails>,
    pub dimensions: Option<String>,
    pub created: Option<String>,
    pub title: Option<String>,
    pub profile_id: Option<String>,
    pub end_date: Option<String>,
    pub self_link: Option<String>,
}

JSON template for Analytics unsampled report resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

unsampled reports insert management (request|response)

unsampled reports get management (response)

Fields
status: Option<String>

Status of this unsampled report. Possible values are PENDING, COMPLETED, or FAILED.

updated: Option<String>

Time this unsampled report was last modified.

download_type: Option<String>

The type of download you need to use for the report data file. Possible values include GOOGLE_DRIVE and GOOGLE_CLOUD_STORAGE. If the value is GOOGLE_DRIVE, see the driveDownloadDetails field. If the value is GOOGLE_CLOUD_STORAGE, see the cloudStorageDownloadDetails field.

start_date: Option<String>

The start date for the unsampled report.

drive_download_details: Option<UnsampledReportDriveDownloadDetails>

Download details for a file stored in Google Drive.

metrics: Option<String>

The metrics for the unsampled report.

filters: Option<String>

The filters for the unsampled report.

web_property_id: Option<String>

Web property ID to which this unsampled report belongs. The web property ID is of the form UA-XXXXX-YY.

segment: Option<String>

The segment for the unsampled report.

id: Option<String>

Unsampled report ID.

account_id: Option<String>

Account ID to which this unsampled report belongs.

kind: Option<String>

Resource type for an Analytics unsampled report.

cloud_storage_download_details: Option<UnsampledReportCloudStorageDownloadDetails>

Download details for a file stored in Google Cloud Storage.

dimensions: Option<String>

The dimensions for the unsampled report.

created: Option<String>

Time this unsampled report was created.

title: Option<String>

Title of the unsampled report.

profile_id: Option<String>

View (Profile) ID to which this unsampled report belongs.

end_date: Option<String>

The end date for the unsampled report.

self_link: Option<String>

Link for this unsampled report.