Struct google_analytics3::Webproperty

pub struct Webproperty {
    pub updated: Option<String>,
    pub default_profile_id: Option<String>,
    pub profile_count: Option<i32>,
    pub internal_web_property_id: Option<String>,
    pub child_link: Option<WebpropertyChildLink>,
    pub industry_vertical: Option<String>,
    pub id: Option<String>,
    pub account_id: Option<String>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub created: Option<String>,
    pub level: Option<String>,
    pub website_url: Option<String>,
    pub parent_link: Option<WebpropertyParentLink>,
    pub permissions: Option<WebpropertyPermissions>,
    pub self_link: Option<String>,
}

JSON template for an Analytics web property.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

webproperties insert management (request|response)

webproperties update management (request|response)

webproperties get management (response)

webproperties patch management (request|response)

Fields

updated

Time this web property was last modified.

default_profile_id

Default view (profile) ID.

profile_count

View (Profile) count for this web property.

internal_web_property_id

Internal ID for this web property.

child_link

Child link for this web property. Points to the list of views (profiles) for this web property.

industry_vertical

The industry vertical/category selected for this web property.

id

Web property ID of the form UA-XXXXX-YY.

account_id

Account ID to which this web property belongs.

kind

Resource type for Analytics WebProperty.

name

Name of this web property.

created

Time this web property was created.

level

Level for this web property. Possible values are STANDARD or PREMIUM.

website_url

Website url for this web property.

parent_link

Parent link for this web property. Points to the account to which this web property belongs.

permissions

Permissions the user has for this web property.

self_link

Link for this web property.